requirements.txt
pandas
streamlit
openpyxl
Matplotlib
Plotly
request